python/requirements.txt
configparser==3.5.0
imutils==0.4.6
matplotlib==2.2.2
numpy==1.14.2
opencv-python==3.4.0.12
python-interface==1.4.0